+++ app/src/main/java/org/gammf/collabora_android/app/gui/module/EditModuleFragment.java
@@ -1,4 +1,4 @@
-package org.gammf.collabora_android.app.gui;
+package org.gammf.collabora_android.app.gui.module;
 
 
 import android.os.Bundle;
