+++ app/src/main/java/org/gammf/collabora_android/app/gui/note/CreateNoteFragment.java
@@ -1,4 +1,4 @@
-package org.gammf.collabora_android.app.gui;
+package org.gammf.collabora_android.app.gui.note;
 
 import android.app.DatePickerDialog;
 import android.app.TimePickerDialog;
