+++ app/src/main/res/layout/fragment_create_module.xml
@@ -3,7 +3,7 @@
              android:layout_width="match_parent"
              android:layout_height="match_parent"
              xmlns:app="http://schemas.android.com/apk/res-auto"
-             tools:context="org.gammf.collabora_android.app.gui.CreateModuleFragment">
+             tools:context="org.gammf.collabora_android.app.gui.module.CreateModuleFragment">
 
     <ImageView
         android:id="@+id/imageView14"
