+++ app/src/main/res/layout/fragment_create_note.xml
@@ -3,7 +3,7 @@
                 xmlns:tools="http://schemas.android.com/tools"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
-                tools:context="org.gammf.collabora_android.app.gui.CreateNoteFragment">
+                tools:context="org.gammf.collabora_android.app.gui.note.CreateNoteFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <EditText
