+++ app/src/main/res/layout/fragment_dialog_new_collaboration.xml
@@ -4,7 +4,7 @@
              android:layout_height="match_parent"
              android:layout_margin="10dp"
              xmlns:app="http://schemas.android.com/apk/res-auto"
-             tools:context="org.gammf.collabora_android.app.gui.DialogNewCollaborationFragment">
+             tools:context="org.gammf.collabora_android.app.gui.collaboration.DialogNewCollaborationFragment">
 
     <ImageView
         android:id="@+id/imageView9"
