+++ app/src/main/res/layout/fragment_edit_collaboration.xml
@@ -3,7 +3,7 @@
              xmlns:tools="http://schemas.android.com/tools"
              android:layout_width="match_parent"
              android:layout_height="match_parent"
-             tools:context="org.gammf.collabora_android.app.gui.EditCollaborationFragment">
+             tools:context="org.gammf.collabora_android.app.gui.collaboration.EditCollaborationFragment">
 
     <TextView
         android:id="@+id/textView15"
